test(BA-5983): consolidate merge tests into the DB-backed file
The synthetic merge test file (test_model_definition_merge.py)
duplicated every scenario the DB-backed test already covers — same
"empty input + baseline → resolved values" and "missing field
→ raise" shapes, but with hand-built drafts instead of the real
``RevisionDraftReader`` path. Keep the realistic version
(test_revision_merge_db.py) as the single source of truth for the
merge contract.

The one piece worth preserving from the synthetic file is the pure
``to_draft`` conversion check. Move that into the existing DTO test
module as ``TestModelDefinitionInputToDraft``, tightened to assert
``draft.model_fields_set == input.model_fields_set`` — the precise
invariant the merge logic relies on (BA-5983 broke because
``model_dump()`` was used without ``exclude_unset=True``, leaving
every field "set" on the draft).
